api-server/requirements.txt
Flask-SQLAlchemy==3.1.1
flask==3.0.2
gunicorn==21.2.0 
openai==1.58.1
psycopg2==2.9.9
python-dotenv==1.0.1